feat(controller): 统一集群ID参数类型为Long，提升数据一致性
- 将ClusterInfoController和NoticeGroupController中的集群ID参数类型从Integer更改为Long，以确保与数据模型的一致性。
- 更新相关方法的参数类型，确保系统在处理集群ID时的稳定性与兼容性。

这些改动显著增强了API的健壮性与数据一致性，确保了整体设计的一致性。

+++ datasophon-api/src/main/java/com/datasophon/api/controller/v1/cluster/ClusterInfoController.java
@@ -85,7 +85,7 @@ public class ClusterInfoController {
      * 信息
      */
     @RequestMapping("/info/{id}")
-    public Result<ClusterInfoVO> info(@PathVariable("id") Integer id) {
+    public Result<ClusterInfoVO> info(@PathVariable("id") Long id) {
         try {
             ClusterInfoDTO clusterInfoDTO = clusterInfoService.getClusterById(id);
             ClusterInfoVO clusterInfoVO = clusterInfoConverter.dtoToVo(clusterInfoDTO);
